Finishes Aurora SkyFishing api + AFSMapper bug fix for missing files.
Finishes Aurora SkyFishing api + AFSMapper bug fix for missing files.

+++ BackendServices/AuxiliaryServices/HomeTools/AFS/AFSClass.cs
@@ -40,7 +40,7 @@ namespace HomeTools.AFS
                         if (!string.IsNullOrEmpty(destinationDirectory) && !Directory.Exists(destinationDirectory))
                             Directory.CreateDirectory(destinationDirectory.ToUpper());
 
-                        if (!File.Exists(NewfilePath))
+                        if (File.Exists(filePath) && !File.Exists(NewfilePath))
                             File.Move(filePath, NewfilePath.ToUpper());
 
                         if (File.Exists(NewfilePath))
@@ -83,7 +83,7 @@ namespace HomeTools.AFS
                                         if (!string.IsNullOrEmpty(destinationDirectory) && !Directory.Exists(destinationDirectory))
                                             Directory.CreateDirectory(destinationDirectory.ToUpper());
 
-                                        if (!File.Exists(NewfilePath))
+                                        if (File.Exists(filePath) && !File.Exists(NewfilePath))
                                             File.Move(filePath, NewfilePath.ToUpper());
 
                                         if (File.Exists(NewfilePath))
@@ -114,7 +114,7 @@ namespace HomeTools.AFS
                             if (!string.IsNullOrEmpty(destinationDirectory) && !Directory.Exists(destinationDirectory))
                                 Directory.CreateDirectory(destinationDirectory.ToUpper());
 
-                            if (!File.Exists(NewfilePath))
+                            if (File.Exists(filePath) && !File.Exists(NewfilePath))
                                 File.Move(filePath, NewfilePath.ToUpper());
 
                             if (File.Exists(NewfilePath))
